feat: update `upload|download_factor` to `i16` instead of `f64`

+++ backend/api/src/env.rs
@@ -62,9 +62,9 @@ pub struct TrackerConfig {
     #[envconfig(from = "ARCADIA_ALLOWED_TORRENT_CLIENTS")]
     pub allowed_torrent_clients: AllowedTorrentClientSet,
     #[envconfig(from = "ARCADIA_GLOBAL_UPLOAD_FACTOR")]
-    pub global_upload_factor: f64,
+    pub global_upload_factor: i16,
     #[envconfig(from = "ARCADIA_GLOBAL_DOWNLOAD_FACTOR")]
-    pub global_download_factor: f64,
+    pub global_download_factor: i16,
     #[envconfig(from = "TASK_INTERVAL_UPDATE_TORRENT_SEEDERS_LEECHERS")]
     pub interval_update_torrent_seeders_leechers: String,
     #[envconfig(from = "TASK_INTERVAL_REMOVE_INACTIVE_PEERS")]
